+++ pkg/descheduler/strategies/duplicates_test.go
@@ -199,9 +199,10 @@ func TestFindDuplicatePods(t *testing.T) {
 			false,
 			testCase.maxPodsToEvict,
 			[]*v1.Node{node},
+			false,
 		)
 
-		RemoveDuplicatePods(ctx, fakeClient, testCase.strategy, []*v1.Node{node}, false, podEvictor)
+		RemoveDuplicatePods(ctx, fakeClient, testCase.strategy, []*v1.Node{node}, podEvictor)
 		podsEvicted := podEvictor.TotalEvicted()
 		if podsEvicted != testCase.expectedEvictedPodCount {
 			t.Errorf("Test error for description: %s. Expected evicted pods count %v, got %v", testCase.description, testCase.expectedEvictedPodCount, podsEvicted)
